requirements.txt

Flask==3.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
openpyxl==3.1.2
watchdog==3.0.0